+++ backend/voicepassing/src/main/java/com/ssafy/voicepassing/model/entity/Result.java
@@ -25,7 +25,7 @@ public class Result extends BaseTimeEntity {
     private int category;
 
     @Column(nullable = false)
-    private int risk;
+    private float risk;
 
 
 }
